move dnd functions to custom hook
